Split out ems daemon, client api and paper engine into new mods
